Add files via upload
Each show the stages of the almost ready: Overly Compensated Watch Glove, that consists of a glove, 3 breadboards, LEDs, a DateandTime Module, 3 LCD screens, an On and Off power switch an Ultrasound Module, a bunch of buttons, a bunch of wire, and a bunch of resistors. Oh, and a 2650 Mega Arduino Board. I recommend trying to make it, if you can. It Is Worth Spending :)